% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{evergreen}
\alias{evergreen}
\title{evergreen palette generator}
\usage{
evergreen(n)
}
\arguments{
\item{n}{Number of colors desired.}
}
\value{
A vector of colors.
}
\description{
evergreen palette generator
}
\examples{
evergreen(3)
}
