% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{evergreen_palettes}
\alias{evergreen_palettes}
\title{evergreen_palettes palette}
\format{
An object of class \code{list} of length 1.
}
\usage{
evergreen_palettes
}
\description{
evergreen_palettes palette
}
\keyword{datasets}
